add serialization checks to test suite (#2190)
