+++ server/util.js
@@ -31,7 +31,34 @@ function wrapAsync(fn) {
     };
 }
 
+function detectBinary(state, path = '', results = []) {
+    const allowedTypes = ['Array', 'Boolean', 'Date', 'Number', 'Object', 'String'];
+
+    if(!state) {
+        return results;
+    }
+
+    let type = state.constructor.name;
+
+    if(!allowedTypes.includes(type)) {
+        results.push({ path: path, type: type });
+    }
+
+    if(type === 'Object') {
+        for(let key in state) {
+            detectBinary(state[key], `${path}.${key}`, results);
+        }
+    } else if(type === 'Array') {
+        for(let i = 0; i < state.length; ++i) {
+            detectBinary(state[i], `${path}[${i}]`, results);
+        }
+    }
+
+    return results;
+}
+
 module.exports = {
+    detectBinary: detectBinary,
     escapeRegex: escapeRegex,
     httpRequest: httpRequest,
     wrapAsync: wrapAsync
