+++ test/helpers/playerinteractionwrapper.js
@@ -1,6 +1,7 @@
 const _ = require('underscore');
 
 const { matchCardByNameAndPack } = require('./cardutil.js');
+const { detectBinary } = require('../../server/util');
 
 class PlayerInteractionWrapper {
     constructor(game, player) {
@@ -442,6 +443,7 @@ class PlayerInteractionWrapper {
 
         this.game.menuButton(this.player.name, promptButton.arg, promptButton.uuid, promptButton.method);
         this.game.continue();
+        this.checkUnserializableGameState();
     }
 
     clickCard(card, location = 'any', side) {
@@ -450,12 +452,14 @@ class PlayerInteractionWrapper {
         }
         this.game.cardClicked(this.player.name, card.uuid);
         this.game.continue();
+        this.checkUnserializableGameState();
         return card;
     }
 
     clickRing(element) {
         this.game.ringClicked(this.player.name, element);
         this.game.continue();
+        this.checkUnserializableGameState();
     }
 
     clickMenu(card, menuText) {
@@ -471,11 +475,13 @@ class PlayerInteractionWrapper {
 
         this.game.menuItemClick(this.player.name, card.uuid, items[0]);
         this.game.continue();
+        this.checkUnserializableGameState();
     }
 
     dragCard(card, targetLocation) {
         this.game.drop(this.player.name, card.uuid, card.location, targetLocation);
         this.game.continue();
+        this.checkUnserializableGameState();
     }
 
     /**
@@ -710,6 +716,15 @@ class PlayerInteractionWrapper {
             }
             return !card.hasDash(type);
         });
+    }
+
+    checkUnserializableGameState() {
+        let state = this.game.getState(this.player.name);
+        let results = detectBinary(state);
+
+        if(results.length !== 0) {
+            throw new Error('Unable to serialize game state back to client:\n' + JSON.stringify(results));
+        }
     }
 }
